+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Add followers endpoints [#58](https://github.com/plotwist-app/plotwist-backend/pull/58/)
+
 ### Added
 
 - AWS SQS setup [(#57)](https://github.com/plotwist-app/plotwist-backend/pull/57)
@@ -12,7 +14,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Organized project envs to use as config in the project [(#57)](https://github.com/plotwist-app/plotwist-backend/pull/57)
-
 
 ### 1.2.2